+++ quantify_scheduler/backends/qblox/helpers.py
@@ -2,13 +2,21 @@
 # Licensed according to the LICENCE file on the main branch
 """Helper functions for Qblox backend."""
 
+import re
+from copy import deepcopy
 from collections import UserDict
 from typing import Any, Dict, Iterable, List, Literal, Tuple, Union
 
 import numpy as np
 
+from quantify_core.utilities.general import without
+
 from quantify_scheduler.backends.qblox import constants
 from quantify_scheduler.helpers.waveforms import exec_waveform_function
+from quantify_scheduler import Schedule
+
+from quantify_scheduler.backends.types.qblox import OpInfo
+from quantify_scheduler.operations.pulse_library import WindowOperation
 
 
 # pylint: disable=invalid-name
@@ -395,3 +403,197 @@ def get_nco_phase_arguments(phase_deg: float) -> Tuple[int, int, int]:
     phase_ultra_fine = int(remaining_phase // constants.NCO_PHASE_DEG_STEP_U_FINE)
 
     return phase_coarse, phase_fine, phase_ultra_fine
+
+
+def generate_port_clock_to_device_map(
+    hardware_cfg: Dict[str, Any]
+) -> Dict[Tuple[str, str], str]:
+    """
+    Generates a mapping that specifies which port-clock combinations belong to which
+    device.
+
+    .. note::
+        The same device may contain multiple port-clock combinations, but each
+        port-clock combination may only occur once.
+
+    Parameters
+    ----------
+    hardware_cfg:
+        The hardware config dictionary.
+
+    Returns
+    -------
+    :
+        A dictionary with as key a tuple representing a port-clock combination, and
+        as value the name of the device. Note that multiple port-clocks may point to
+        the same device.
+    """
+
+    portclock_map = dict()
+    for device_name, device_info in hardware_cfg.items():
+        if not isinstance(device_info, dict):
+            continue
+
+        portclocks = find_all_port_clock_combinations(device_info)
+
+        for portclock in portclocks:
+            portclock_map[portclock] = device_name
+
+    return portclock_map
+
+
+# pylint: disable=too-many-locals
+# pylint: disable=too-many-branches
+def assign_pulse_and_acq_info_to_devices(
+    schedule: Schedule,
+    device_compilers: Dict[str, Any],
+    hardware_cfg: Dict[str, Any],
+):
+    """
+    Traverses the schedule and generates `OpInfo` objects for every pulse and
+    acquisition, and assigns it to the correct `InstrumentCompiler`.
+
+    Parameters
+    ----------
+    schedule
+        The schedule to extract the pulse and acquisition info from.
+    device_compilers
+        Dictionary containing InstrumentCompilers as values and their names as keys.
+    hardware_cfg
+        The hardware config dictionary.
+
+    Raises
+    ------
+    RuntimeError
+        This exception is raised then the function encountered an operation that has no
+        pulse or acquisition info assigned to it.
+    KeyError
+        This exception is raised when attempting to assign a pulse with a port-clock
+        combination that is not defined in the hardware configuration.
+    KeyError
+        This exception is raised when attempting to assign an acquisition with a
+        port-clock combination that is not defined in the hardware configuration.
+    """
+
+    portclock_mapping = generate_port_clock_to_device_map(hardware_cfg)
+
+    for op_timing_constraint in schedule.schedulables.values():
+        op_hash = op_timing_constraint["operation_repr"]
+        op_data = schedule.operations[op_hash]
+
+        if isinstance(op_data, WindowOperation):
+            continue
+
+        if not op_data.valid_pulse and not op_data.valid_acquisition:
+            raise RuntimeError(
+                f"Operation {op_hash} is not a valid pulse or acquisition. Please check"
+                f" whether the device compilation been performed successfully. "
+                f"Operation data: {repr(op_data)}"
+            )
+
+        operation_start_time = op_timing_constraint["abs_time"]
+        for pulse_data in op_data.data["pulse_info"]:
+            if "t0" in pulse_data:
+                pulse_start_time = operation_start_time + pulse_data["t0"]
+            else:
+                pulse_start_time = operation_start_time
+
+            port = pulse_data["port"]
+            clock = pulse_data["clock"]
+            if port is None:
+                continue  # ignore idle pulses
+
+            combined_data = OpInfo(
+                name=op_data.data["name"],
+                data=pulse_data,
+                timing=pulse_start_time,
+            )
+
+            if (port, clock) not in portclock_mapping:
+                raise KeyError(
+                    f"Could not assign pulse data to device. The combination"
+                    f" of port {port} and clock {clock} could not be found "
+                    f"in hardware configuration.\n\nAre both the port and clock "
+                    f"specified in the hardware configuration?\n\nRelevant operation:\n"
+                    f"{combined_data}."
+                )
+            dev = portclock_mapping[(port, clock)]
+            device_compilers[dev].add_pulse(port, clock, pulse_info=combined_data)
+
+        for acq_data in op_data.data["acquisition_info"]:
+            if "t0" in acq_data:
+                acq_start_time = operation_start_time + acq_data["t0"]
+            else:
+                acq_start_time = operation_start_time
+            port = acq_data["port"]
+            clock = acq_data["clock"]
+            if port is None:
+                continue
+
+            hashed_dict = without(acq_data, ["t0", "waveforms"])
+            hashed_dict["waveforms"] = list()
+            for acq in acq_data["waveforms"]:
+                hashed_dict["waveforms"].append(without(acq, ["t0"]))
+
+            combined_data = OpInfo(
+                name=op_data.data["name"],
+                data=acq_data,
+                timing=acq_start_time,
+            )
+            if (port, clock) not in portclock_mapping:
+                raise KeyError(
+                    f"Could not assign acquisition data to device. The combination"
+                    f" of port {port} and clock {clock} could not be found "
+                    f"in hardware configuration.\n\nAre both the port and clock "
+                    f"specified in the hardware configuration?\n\nRelevant operation:\n"
+                    f"{combined_data}."
+                )
+            dev = portclock_mapping[(port, clock)]
+            device_compilers[dev].add_acquisition(port, clock, acq_info=combined_data)
+
+
+def convert_hw_config_to_portclock_configs_spec(
+    hw_config: Dict[str, Any],
+):
+    """
+    Converts possibly old hardware configs to the new format introduced by
+    the new dynamic sequencer allocation feature. I.e. manual
+    assignment between sequencers and portclocks under each output is removed, and
+    instead only a list of port-clock configurations is specified,
+    under the new `portclock_configs` key.
+
+    Parameters
+    ----------
+    hw_config
+        The hardware config to be upgraded to the new specification.
+
+    Returns
+    -------
+    :
+        A hardware config compatible with the specification required by the new
+        dynamic sequencer allocation feature.
+
+    """
+
+    hw_config = deepcopy(hw_config)
+
+    for device_info in hw_config.values():
+        if not isinstance(device_info, dict):
+            continue
+
+        for io, io_cfg in device_info.items():
+            if not isinstance(io_cfg, dict):
+                continue
+
+            new_io_cfg = {}
+            new_io_cfg["portclock_configs"] = io_cfg.get("portclock_configs", [])
+
+            for entry, value in io_cfg.items():
+                if not re.match("^seq\d+$", entry):
+                    new_io_cfg[entry] = value
+                    continue
+                new_io_cfg["portclock_configs"].append(io_cfg[entry])
+
+            device_info[io] = new_io_cfg
+
+    return hw_config