+++ tests/fixtures/schedule.py
@@ -9,7 +9,7 @@
 from __future__ import annotations
 
 from copy import deepcopy
-from typing import Any, Callable, Dict, Optional
+from typing import Any, Dict, Optional, Callable, List
 
 import numpy as np
 import pytest
@@ -108,6 +108,17 @@ def make_basic_schedule() -> Callable[[str], Schedule]:
     def _make_basic_schedule(qubit: str) -> Schedule:
         schedule = Schedule(f'Basic schedule{" "+qubit if qubit != "q0" else ""}')
         schedule.add(X90(qubit))
+        return schedule
+
+    return _make_basic_schedule
+
+
+@pytest.fixture
+def make_basic_multi_qubit_schedule() -> Callable[[List[str]], Schedule]:
+    def _make_basic_schedule(qubits: List[str]) -> Schedule:
+        schedule = Schedule(f"Basic schedule {qubits}")
+        for qubit in qubits:
+            schedule.add(X90(qubit))
         return schedule
 
     return _make_basic_schedule